+++ apps/desktop/src/app/vault/vault-filter/filters/collection-filter.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 
-import { CollectionFilterComponent as BaseCollectionFilterComponent } from "@bitwarden/angular/modules/vault-filter/components/collection-filter.component";
+import { CollectionFilterComponent as BaseCollectionFilterComponent } from "@bitwarden/angular/vault/vault-filter/components/collection-filter.component";
 
 @Component({
   selector: "app-collection-filter",
