+++ apps/desktop/src/app/vault/vault-filter/filters/organization-filter.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 
-import { OrganizationFilterComponent as BaseOrganizationFilterComponent } from "@bitwarden/angular/modules/vault-filter/components/organization-filter.component";
-import { DisplayMode } from "@bitwarden/angular/modules/vault-filter/models/display-mode";
+import { OrganizationFilterComponent as BaseOrganizationFilterComponent } from "@bitwarden/angular/vault/vault-filter/components/organization-filter.component";
+import { DisplayMode } from "@bitwarden/angular/vault/vault-filter/models/display-mode";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
 import { Organization } from "@bitwarden/common/models/domain/organization";
