+++ apps/desktop/src/app/vault/vault-filter/filters/status-filter.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 
-import { StatusFilterComponent as BaseStatusFilterComponent } from "@bitwarden/angular/modules/vault-filter/components/status-filter.component";
+import { StatusFilterComponent as BaseStatusFilterComponent } from "@bitwarden/angular/vault/vault-filter/components/status-filter.component";
 
 @Component({
   selector: "app-status-filter",
