+++ apps/desktop/src/app/vault/vault-filter/filters/type-filter.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 
-import { TypeFilterComponent as BaseTypeFilterComponent } from "@bitwarden/angular/modules/vault-filter/components/type-filter.component";
+import { TypeFilterComponent as BaseTypeFilterComponent } from "@bitwarden/angular/vault/vault-filter/components/type-filter.component";
 
 @Component({
   selector: "app-type-filter",
