+++ apps/desktop/src/app/vault/vault-filter/vault-filter.component.ts
@@ -1,6 +1,6 @@
 import { Component } from "@angular/core";
 
-import { VaultFilterComponent as BaseVaultFilterComponent } from "@bitwarden/angular/modules/vault-filter/vault-filter.component";
+import { VaultFilterComponent as BaseVaultFilterComponent } from "@bitwarden/angular/vault/vault-filter/components/vault-filter.component";
 
 @Component({
   selector: "app-vault-filter",
