+++ apps/desktop/src/app/vault/vault-filter/vault-filter.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 
 import { JslibModule } from "@bitwarden/angular/jslib.module";
-import { VaultFilterService } from "@bitwarden/angular/modules/vault-filter/vault-filter.service";
+import { VaultFilterService } from "@bitwarden/angular/vault/vault-filter/services/vault-filter.service";
 
 import { CollectionFilterComponent } from "./filters/collection-filter.component";
 import { FolderFilterComponent } from "./filters/folder-filter.component";
