+++ apps/desktop/src/app/vault/vault.component.ts
@@ -11,8 +11,8 @@ import { ActivatedRoute, Router } from "@angular/router";
 import { first } from "rxjs/operators";
 
 import { ModalRef } from "@bitwarden/angular/components/modal/modal.ref";
-import { VaultFilter } from "@bitwarden/angular/modules/vault-filter/models/vault-filter.model";
 import { ModalService } from "@bitwarden/angular/services/modal.service";
+import { VaultFilter } from "@bitwarden/angular/vault/vault-filter/models/vault-filter.model";
 import { BroadcasterService } from "@bitwarden/common/abstractions/broadcaster.service";
 import { EventService } from "@bitwarden/common/abstractions/event.service";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
