+++ apps/web/src/app/organizations/vault/organization-vault.component.ts
@@ -10,8 +10,8 @@ import {
 import { ActivatedRoute, Params, Router } from "@angular/router";
 import { first } from "rxjs/operators";
 
-import { VaultFilter } from "@bitwarden/angular/modules/vault-filter/models/vault-filter.model";
 import { ModalService } from "@bitwarden/angular/services/modal.service";
+import { VaultFilter } from "@bitwarden/angular/vault/vault-filter/models/vault-filter.model";
 import { BroadcasterService } from "@bitwarden/common/abstractions/broadcaster.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
