+++ apps/web/src/app/vault/shared/vault.service.ts
@@ -1,4 +1,4 @@
-import { VaultFilter } from "@bitwarden/angular/modules/vault-filter/models/vault-filter.model";
+import { VaultFilter } from "@bitwarden/angular/vault/vault-filter/models/vault-filter.model";
 
 export class VaultService {
   calculateSearchBarLocalizationString(vaultFilter: VaultFilter): string {
