+++ libs/angular/src/vault/vault-filter/components/collection-filter.component.ts
