+++ libs/angular/src/vault/vault-filter/components/folder-filter.component.ts
