+++ libs/angular/src/vault/vault-filter/models/display-mode.ts
