+++ libs/angular/src/vault/vault-filter/models/dynamic-tree-node.model.ts
