+++ libs/angular/src/vault/vault-filter/services/vault-filter.service.ts
@@ -14,7 +14,7 @@ import { TreeNode } from "@bitwarden/common/models/domain/treeNode";
 import { CollectionView } from "@bitwarden/common/models/view/collectionView";
 import { FolderView } from "@bitwarden/common/models/view/folderView";
 
-import { DynamicTreeNode } from "./models/dynamic-tree-node.model";
+import { DynamicTreeNode } from "../models/dynamic-tree-node.model";
 
 const NestingDelimiter = "/";
